requirements.txt
jpype1==1.3.0
opencv-python-headless
